requirements.txt
# Core dependencies
fastmcp
reflex
streamlit
python-dotenv
openai
fastapi
uvicorn[standard]
pydantic-settings
Jinja2
httpx

# Geo-location & mapping
geopy
folium
streamlit-folium

# Additional utilities
requests
